refactor(#13): 집 찾기 뷰모델 위치 변경
